Forgot to push last edit, heaps of improvements and fleshing out
The first of the two (maybe three/four) buildings is completed as a static mesh. UVs fixed via Blender.

Tons of improvement regarding building sizing, placement, and scale. Cliff in the water are also being refined a ton. Scale was a little wonky on the cliff islands but it's been revised since.

Meant to push out an update before this but failed to push the commit, hence the large amount of improvement.